Create main class to auto calculate weights

+++ src/utils/Global.java
@@ -1,14 +1,8 @@
 package utils;
 
-import repast.simphony.random.RandomHelper;
-import repast.simphony.space.SpatialMath;
-import repast.simphony.space.continuous.ContinuousSpace;
-import repast.simphony.space.continuous.NdPoint;
-import repast.simphony.space.grid.Grid;
-
 public class Global {
 
-	public static float RATIO = 8;
+	public static int RATIO = 8;
 	public static double KILL_CHANCE = 0.6;
 
 	public static int dormants = 0;
